Add files via upload
This is a new Website built which is responsive for mobile, tablet and computers. It uses JS, CSS and HTML for various forms, add and deleting new items to the basket, button such as Print, Submit and Reset. Please Attempt to fill out the forms to show I have used JS basic concepts. I have also used Google Maps embedded and other functions. 

This is a sample website which I have used to demonstrate my skills and knowledge as a front end developer.